+++ src/main/java/warehouse/LogicDTOs/RackDTO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class RackDTO {
     private Long rackId;
-    private Long area;
-    private Long capacity;
+    private Double area;
+    private Double capacity;
 
     // Getters and Setters
 }
